requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
slowapi==0.1.9
groq==0.23.1
httpx==0.28.1
opencv-python-headless==4.11.0.86
Pillow==11.2.1
numpy==2.2.5
pytesseract==0.3.13
pymupdf==1.24.14
pydantic==2.10.6
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
pdf2image==1.17.0
pytest==8.3.5
pytest-asyncio==0.26.0
